requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy<2.0
open-clip-torch